+++ CHANGELOG.md
@@ -16,6 +16,9 @@ _**Note:** Yet to be released breaking changes appear here._
   TypeScript users must update their code to handle the `undefined` case when calling this method.
 - The `Dictionary` class has been removed. The maxGraph API use the `Map` class instead, which is a standard JavaScript feature.
   If your code depends on the `Dictionary` class, you can use the `Map` class instead.
+- The return types of some methods of EditorToolbar are now more precise.
+  - addPrototype(): HTMLImageElement, no longer HTMLImageElement | HTMLButtonElement
+  - addCombo(): HTMLSelectElement instead of HTMLElement
 
 ## 0.20.0